sleepy boy

+++ src/hooks/Pause.cpp
@@ -14,8 +14,8 @@ class $modify(RandomPausePlayerObject, PlayerObject) {
         if (auto playLayer = PlayLayer::get()) {
             if (horribleMod->getSavedValue<bool>("pauses", false)) {
                 auto rnd = rand() % 101;
-
-                if (rnd <= static_cast<int>(horribleMod->getSettingValue<int64_t>("pauses-chance"))) {
+                // if the rng is lower than the chance, pause the game
+                if (rnd <= horribleMod->getSettingValue<int>("pauses-chance")) {
                     log::debug("Pausing the game randomly");
                     playLayer->pauseGame(true); // pause the game randomly
                 };
